Initial Commit for bg_remove_stream testcode
This is a proof of concept to use imxvideoconvert_pxp and opencv.

The Code Takes a FullHD frame and converts it to 360 Pixel (SD):
640 x 360.
From the resized Image, the Color-Format is changed from BGR to HSV.
With the HSV color format it is possible to highlight certain color
spectra to create a mask that hides everything except a PCB.

This step is required to detect a board in an image and resize the image
to focus on the board.

The Preproccessed Image should be used as Input for Image-Classification