Include contours and minarea-Rectang detection
Now we get a rectangle bounding-box around our PCB.

We use our Threshhold-mask to detect the biggest contour.
It is most likly, that the biggest contour is our PCB.

We use minAreaRect to detect the smallest rectangle around the biggest
contour (PCB).

ToDo: get weight and rotation of the PCB and focus board.