Improve Image-Preprocessing and Distortion
Since the camera creates a distorted image through the optics, the image
is rectified. This process is very power intensive.
Since it is assumed that the edge area of the image is less meaningful
for the AI classification due to the angle and the rectification, the
image area is limited to 4:3. This also requires 25% fewer pixels to be
processed and generates better CPU performance.