Resize object image to 320x320x3
To avoid further scaling of the image for Tensorflow, the object image
is saved as a 320x320 image. Undefined areas are defined as black
(R=0;G=0;B=0).

This also has the advantage that the image size is always known and thus
the necessary allocated memory for shared memory can be better
calculated for multi-process tasks.