Implement Multiprocessing for Board classification and dataset creation
The evaluation of object_images can now be done in separate processes.

For this a buffer is created, which contains the image and can be served
by both processes.

The process for the classification can be controlled via a queue.
Signals can be exchanged between processes to perform certain actions,
such as creating an image for the dataset.