extend python multiprocessing
Here are some main improvements done for Python multiprocessing.

Now we have 3 main Processes, each for a specific puprose.
- THT-Classificator.main()
- mp.process_classification()
- mp.process_preprocess()

The "main" process is THT-Classificator.main()
This works as a programm manager, that controles all other processes.

It defines a share-memory buffer used by mp.process_classification() and
mp.process_preprocess()
and manages Signals from and to those Processes.
THT-Classificator.main() also manages a thread that run the sshkeyboard
application to detect a specific key-press.

mp.process_preprocess() does the image-preprocessing. It controles the
ingoing and outgoing videostream and detects a PCB object within every
frame. To keep the preprocess simple, mp.process_preprocess() trys to
align the object, that can be accessed throw a shared buffer, managed by
main().

mp.process_classification() is a initial implementation to handle a
object image comming from a shared buffer. Currently it is used to
handle the dataset creation, but it should be handle the complete
KI-Application in the future.